Move the test of the version endpoint to TestResponse
